backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Vector Database
qdrant-client==1.9.0

# Configuration & Validation
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Authentication & Security
pyjwt==2.8.0
argon2-cffi==23.1.0

# Observability
prometheus-fastapi-instrumentator==7.0.0
python-json-logger==2.0.7

# HTTP Client (for testing)
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
